Rename views as themes/default
