+++ src/index.php
@@ -204,4 +204,4 @@ switch ( $action ) {
 }
 
 # Views
-require_once( "views/html.php" );
+require_once( "themes/default/html.php" );
