+++ src/themes/default/actions/status.php
@@ -1,0 +1,4 @@
+<?= page_title(); ?>
+<?= page_intro(); ?>
+
+<?php require ROOT_DIR . "/themes/default/_items.php"; ?>
