--- src/views/actions/search.php
@@ -1,4 +1,0 @@
-<?= page_title(); ?>
-<?= page_intro(); ?>
-
-<?php require ROOT_DIR . "/views/_items.php"; ?>
